Fixed local authentication strategy,
The username field was always undefined for me regardless of request
parameters. Possible cause is the platform specific behavior in feathers
API:
https://docs.feathersjs.com/api/authentication/local.html#configuration
"Important: If you want to set the value of usernameField to username in
your configuration file under Windows or running the node process
manager PM2 in Ubuntu/Linux, the value has to be escaped as \\username
(otherwise the username environment variable will be used)." The
suggested workaround didn't work, so I just renamed it to email, which
is what is used as username in this authentication method.