Fixed token refresh logic.
It required authentication, but at the same time was not set up to do
any authentication. Fixed by removing the requirement, since from what I
understand you are supposed to be able to get a new auth token from a
refresh token even if the old one is expired. The native interface does
not even use the old auth token in this case, metaverse dashboard does
however.